# src/gen/java/io/swagger/model/WorkInvalidResult.java
/*
 * PTCRISync webservice
 * This is an interface for PTCRISync library.
 *
 * OpenAPI spec version: 0.0.1-alpha
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import io.swagger.model.Work;
import java.util.ArrayList;
import java.util.List;
import javax.validation.constraints.*;

/**
 * WorkInvalidResult
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2017-02-03T10:28:37.821Z")
public class WorkInvalidResult   {
  @JsonProperty("work")
  private Work work = null;

  @JsonProperty("error")
  private List<String> error = new ArrayList<String>();

  public WorkInvalidResult work(Work work) {
    this.work = work;
    return this;
  }

   /**
   * Get work
   * @return work
  **/
  @JsonProperty("work")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public Work getWork() {
    return work;
  }

  public void setWork(Work work) {
    this.work = work;
  }

  public WorkInvalidResult error(List<String> error) {
    this.error = error;
    return this;
  }

  public WorkInvalidResult addErrorItem(String errorItem) {
    this.error.add(errorItem);
    return this;
  }

   /**
   * Get error
   * @return error
  **/
  @JsonProperty("error")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public List<String> getError() {
    return error;
  }

  public void setError(List<String> error) {
    this.error = error;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    WorkInvalidResult workInvalidResult = (WorkInvalidResult) o;
    return Objects.equals(this.work, workInvalidResult.work) &&
        Objects.equals(this.error, workInvalidResult.error);
  }

  @Override
  public int hashCode() {
    return Objects.hash(work, error);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class WorkInvalidResult {\n");
    
    sb.append("    work: ").append(toIndentedString(work)).append("\n");
    sb.append("    error: ").append(toIndentedString(error)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}